refactor: SourceType enum 클래스 공통 모듈로 이동 및 관련 코드 정리
- SourceType 관련 import 경로 업데이트
- ContentProcessedMessage 관련 로직 확장 함수로 리팩토링
- 불필요한 코드 및 주석 제거
