+++ harvest/src/main/kotlin/com/sandro/realtime/harvest/common/repository/SourceContentRepository.kt
@@ -1,7 +1,7 @@
 package com.sandro.realtime.harvest.common.repository
 
+import com.sandro.realtime.common.domain.SourceType
 import com.sandro.realtime.harvest.common.domain.SourceContent
-import com.sandro.realtime.harvest.common.domain.SourceType
 import org.springframework.data.mongodb.repository.MongoRepository
 import org.springframework.data.mongodb.repository.Query
 import org.springframework.stereotype.Repository
